refactor: extract handler functions that use new model
