+++ src/modules/reporting/jobs/updateBookCompletionProgressHandler.ts
@@ -2,47 +2,36 @@ import { createLogger } from "@/logging";
 import { getDb, kyselyTransaction } from "@/db";
 import { sql } from "kysely";
 import { GlossStateRaw } from "@/modules/translation/types";
-import * as z from "zod";
-import { defineJob } from "@/shared/jobs/JobDefinition";
+import { UpdateBookCompletionProgressJob } from "./UpdateBookCompletionProgressJob";
 
-export const UpdateBookCompletionProgressPayloadSchema = z.object({
-  allLanguages: z.boolean().optional().default(false),
-});
 
-export type UpdateBookCompletionProgressPayload = z.output<
-  typeof UpdateBookCompletionProgressPayloadSchema
->;
+export async function updateBookCompletionProgressHandler(
+  job: UpdateBookCompletionProgressJob,
+) {
+  const logger = createLogger({
+    job: {
+      id: job.id,
+      type: job.type,
+    },
+  });
 
-export const updateBookCompletionProgressJob = defineJob({
-  type: "update_book_completion_progress",
-  payloadSchema: UpdateBookCompletionProgressPayloadSchema,
-  async handler(job) {
-    const logger = createLogger({
-      job: {
-        id: job.id,
-        type: job.type,
-      },
-    });
+  const languages =
+    job.payload.allLanguages ?
+      await findAllLanguages()
+    : await findChangedBooks();
 
-    const languages =
-      job.payload.allLanguages ?
-        await findAllLanguages()
-      : await findChangedBooks();
+  logger.info(`Processing ${languages.length} languages(s)`);
 
-    logger.info(`Processing ${languages.length} languages(s)`);
+  for (const { languageId, books } of languages) {
+    await updateLanguageProgress(languageId, books);
 
-    for (const { languageId, books } of languages) {
-      await updateLanguageProgress(languageId, books);
+    logger.info(`Completed ${languageId}`);
+  }
 
-      logger.info(`Completed ${languageId}`);
-    }
+  logger.info(`Completed: ${languages.length} language(s) processed`);
 
-    logger.info(`Completed: ${languages.length} language(s) processed`);
-
-    return;
-  },
-  timeout: 60 * 15,
-});
+  return;
+}
 
 async function updateLanguageProgress(
   languageId: string,
